Rename find strategy to locator.
