+++ lib/locators/class-name.js
