+++ lib/locators/contains-text.js
