+++ lib/locators/custom-label.js
