+++ lib/locators/exact-text.js
