+++ lib/locators/image.js
