+++ lib/locators/name.js
