+++ lib/locators/node-type.js
