+++ lib/locators/value.js
