+++ lib/scope-strategies/discover-parent.js
@@ -33,7 +33,7 @@ var DiscoverParentContainer = function () {
     function DiscoverParentContainer(config) {
         _classCallCheck(this, DiscoverParentContainer);
 
-        this.findStrategy = config.findStrategy;
+        this.locator = config.locator;
         this.modifiers = config.modifiers || {};
 
         this.modifiers = mergeOptions(this.modifiers, _visible2.default);
@@ -53,17 +53,17 @@ var DiscoverParentContainer = function () {
 
             var parent = context;
 
-            var findStrategy = this.findStrategy;
+            var locator = this.locator;
 
             if (target.modifiers.length > 0) {
                 var modifierNames = target.modifiers.filter(function (m) {
                     return _this.modifiers[m].find;
                 });
-                if (modifierNames.length > 0) findStrategy = this.modifiers[modifierNames[0]].find;
+                if (modifierNames.length > 0) locator = this.modifiers[modifierNames[0]].find;
             }
 
             while (parent && elements.length == 0) {
-                elements = findStrategy(target.label, parent, customLabels);
+                elements = locator(target.label, parent, customLabels);
                 parent = parent.parentNode;
             }
