+++ lib/selector.js
@@ -9,7 +9,7 @@ var _discoverParent = require("./scope-strategies/discover-parent");
 
 var _discoverParent2 = _interopRequireDefault(_discoverParent);
 
-var _default = require("./find-strategies/default");
+var _default = require("./locators/default");
 
 var _default2 = _interopRequireDefault(_default);
 
@@ -47,7 +47,7 @@ function GlanceSelector(options) {
         var data = Parser.parse(reference);
 
         var resolvedLabels = resolveCustomLabels(data, _selector.customLabels, _selector);
-        var elements = _selector.containerStrategyFactory({ findStrategy: _default2.default, modifiers: _selector.modifiers }).search(data, document, 0, resolvedLabels);
+        var elements = _selector.containerStrategyFactory({ locator: _default2.default, modifiers: _selector.modifiers }).search(data, document, 0, resolvedLabels);
 
         if (elements.length === 1) return elements[0];else return elements;
     };
