+++ src/locators/exact-text.js
