+++ src/locators/image.js
