+++ src/locators/node-type.js
