+++ src/locators/value.js
